pyproject.toml
[build-system]
requires = ["setuptools >= 60.0.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sibyls"
description = "Production-ready LLM-powered data labeling, quality diagnostics, and distillation export"
readme = "README.md"
requires-python = ">=3.10, <3.13"
version = '0.0.0'
authors = [{name="Nick Hallmark", email="NHALLMARK3@BLOOMBERG.NET"}]


dependencies = [
    # Core dependencies
    'pydantic>=2',
    'pandas>=2',
    'langchain>=0.1.0',
    'openai>=1.0',
    'requests>=2',
    'loguru>=0.7',
    'jinja2>=3.1',
    'faiss-cpu>=1.7',
    'click>=8.0',
    'markdown>=3.4',
    'pyyaml>=6.0',
    "pydantic-settings>=2.9.1",
    "langchain-openai>=0.3.21",
    "sentence-transformers>=4.1.0",
    "langchain-community>=0.3.24",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    # Phase 1 dependencies
    'scipy>=1.11.0', # Confidence calibration
    'krippendorff>=0.8.1', # Inter-annotator agreement
    'hdbscan>=0.8.33', # Cluster fragmentation analysis in diagnostics
    # Phase 2 dependencies
    'networkx>=3.0', # GraphRAG graph operations
    'python-louvain>=0.16', # GraphRAG community detection
    'umap-learn>=0.5.0', # RAPTOR dimensionality reduction
    'rank-bm25>=0.2.2', # Hybrid search BM25 component
    # Core LLM providers
    'anthropic>=0.39.0',
    'google-generativeai>=0.8.3',
    "dspy-ai>=3.0.3",
]

[project.scripts]
sibyls = "sibyls.cli:cli"

[project.optional-dependencies]
dev = [
    # Standard dev workflow
    'black',
    'build',
    'codespell',
    'pre-commit',
    'pytest',
    'pytest-cov',
    'pytest-benchmark',
    'pytest-asyncio',
    'pytest-mock',
    'pytest-timeout',
    'ruff',
    'twine',
]

jupyter = [
    'ipykernel',
    'jupyter',
]

dashboard = [
    'streamlit>=1.43.0',  # Phase 1 quality dashboard
    'plotly>=5.24.0',  # Interactive visualizations
]

lit = [
    'lit-nlp>=1.2',  # Language Interpretability Tool for distilled model exploration
]

# Phase 2 optional dependencies (require additional setup)
dspy = [
    'dspy-ai>=2.5.0',  # DSPy prompt optimization framework
]

versioning = [
    'dvc>=3.0.0',  # Data version control
    'dvc-s3>=3.0.0',  # S3 remote storage support
    'dvc-azure>=3.0.0',  # Azure Blob Storage support
    'dvc-gs>=3.0.0',  # Google Cloud Storage support
]

# Install all Phase 2 features
phase2 = [
    'dspy-ai>=2.5.0',
    'dvc>=3.0.0',
]

# Probe model: lightweight local fine-tuning for fast training-data iteration
probe = [
    'transformers>=4.46.0',  # Model loading, Trainer, AutoModel
    'torch>=2.0.0',          # PyTorch backend
    'accelerate>=1.2.0',     # Trainer backend utilities
    'datasets>=2.14.0',      # HuggingFace Dataset class used in ProbeTrainer
]

# Phase 3 optional dependencies (DPO/RLHF - requires GPU)
dpo = [
    'transformers>=4.46.0',  # Model loading and training
    'trl>=0.12.1',  # DPO trainer
    'peft>=0.13.0',  # LoRA for efficient fine-tuning
    'accelerate>=1.2.0',  # Distributed training utilities
    'torch>=2.0.0',  # PyTorch backend
]

# Install all Phase 3 features (archived - kept for reference)
phase3 = [
    'langgraph>=0.2.0',  # Multi-agent orchestration (archived)
    'langsmith>=0.2.0',  # Agent monitoring (archived)
    'evidently>=0.5.8',  # Drift detection (archived)
]

# Complete installation with all features
all = [
    'streamlit>=1.43.0',
    'plotly>=5.24.0',
    'lit-nlp>=1.2',
    'dspy-ai>=2.5.0',
    'dvc>=3.0.0',
    'dvc-s3>=3.0.0',
    'dvc-azure>=3.0.0',
    'dvc-gs>=3.0.0',
    'langgraph>=0.2.0',
    'langsmith>=0.2.0',
    'evidently>=0.5.8',
    'transformers>=4.46.0',
    'datasets>=2.14.0',
    'trl>=0.12.1',
    'peft>=0.13.0',
    'accelerate>=1.2.0',
]

[tool.setuptools.packages.find]
where = ["src"]

### Docker ###########################################################################
[tool.docker.base_images.default]
image = ""

[tool.docker.base_images.cuda]
image = ""
cuda_version = "12.1"
cudnn_version = "8.9.2.26"
nccl_version = "2.18.5"

### STANDARDIZED CODE REQUIREMENTS
### black ############################################################################
[tool.black]
line-length = 88 # ensure this matches ruff this is the default
target-version = ['py310', 'py311', 'py312']
include = '\.py$'
skip-string-normalization = true

### pytest ###########################################################################
[tool.pytest.ini_options]
testpaths = [
    "tests",
]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, multiple components)",
    "performance: Performance/benchmark tests",
    "validation: Validation tests on benchmark datasets",
    "slow: Slow-running tests",
    "requires_api: Tests requiring external API access",
]
# Coverage settings
[tool.coverage.run]
source = ["src/sibyls"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true

### ruff #############################################################################
[tool.ruff]
line-length = 88  # ensure this matches black this is the default

[tool.ruff.lint]
select = [
    # All rules enabled as defined here:
    # https://github.com/charliermarsh/ruff#supported-rules
    # Pyflakes
    "F",
    # pydocstyle
    "E",
    "W",
    # isort
    "I",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-unused-arguments
    "ARG",
    # Ruff-specific
    "RUF",
    # pep8-naming
    # InvalidClassName
    "N801",
    # InvalidFunctionName
    "N802",
    # InvalidArgumentName
    "N803",
    # InvalidFirstArgumentNameForClassMethod
    "N804",
    # InvalidFirstArgumentNameForMethod
    "N805",
    # NonLowercaseVariableInFunction
    "N806",
    #pandas-vet
    # UseOfInplaceArgument
    "PD002",
    # UseOfDotIsNull
    "PD003",
    # UseOfDotNotNull
    "PD004",
    # UseOfDotIx
    "PD007",
    # UseOfDotAt
    "PD008",
    # UseOfDotIat
    "PD009",
    # UseOfDotPivotOrUnstack
    "PD010",
    # UseOfDotValues
    "PD011",
    # UseOfDotReadTable
    "PD012",
    # UseOfDotStack
    "PD013",
    # UseOfPdMerge
    "PD015",
]
# Add rules to explicitly ignore here
ignore = [
]
extend-unsafe-fixes = ["UP034"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'

[tool.ruff.format]
quote-style = 'single'

### sqlfluff ########################################################################
[tool.sqlfluff.core]
dialect = "bigquery"

templater = "jinja"

# For a breakdown of rules, check out https://docs.sqlfluff.com/en/stable/rules.html
exclude_rules = "L015,L026,L027,L031,L034,L042,L054"

ignore = "templating"
[tool.sqlfluff.rules]
tab_space_size = 2

[tool.sqlfluff.rules.L010]
capitalisation_policy = "upper"

[tool.sqlfluff.rules.L014]
extended_capitalisation_policy = "lower"
unquoted_identifiers_policy = "all"

[tool.sqlfluff.rules.L016]
ignore_comment_lines = true

[tool.sqlfluff.rules.L030]
extended_capitalisation_policy = "upper"

[tool.codespell]
exclude-file = ".codespellignore"
